Renamed common services
